+++ src/env.ts
@@ -1,5 +1,3 @@
-import type { SupportedChain } from "lib/data";
-
 export const SUPPORTED_CHAIN_IDS: string[] = (() => {
   const chainIds = process.env.NEXT_PUBLIC_SUPPORTED_CHAIN_IDS?.split(",");
   if (!chainIds)
@@ -18,5 +16,4 @@ export const DEFAULT_SUPPORTED_CHAIN_ID = SUPPORTED_CHAIN_IDS[0];
 
 export const DUMMY_MNEMONIC = process.env.NEXT_PUBLIC_DUMMY_MNEMONIC;
 
-export const SELECTED_CHAIN = process.env
-  .NEXT_PUBLIC_SELECTED_CHAIN as SupportedChain;
+export const SELECTED_CHAIN = process.env.NEXT_PUBLIC_SELECTED_CHAIN;