+++ src/lib/app-provider/hooks/index.ts
@@ -10,7 +10,6 @@ export * from "./useRestrictedInput";
 export * from "./useSelectChain";
 export * from "./useSimulateFee";
 export * from "./useTokensInfo";
-export * from "./useCurrentNetwork";
 export * from "./useBaseApiRoute";
 export * from "./useRPCEndpoint";
 export * from "./useFaucetConfig";