+++ src/lib/components/PermissionChip.tsx
@@ -2,9 +2,12 @@ import { chakra, Tag, Tooltip } from "@chakra-ui/react";
 import { useWallet } from "@cosmos-kit/react";
 import type { CSSProperties } from "react";
 
-import type { HumanAddr, PermissionAddresses } from "lib/types";
-import { AccessConfigPermission } from "lib/types";
-import { getPermissionHelper } from "lib/utils";
+import type {
+  HumanAddr,
+  PermissionAddresses,
+  AccessConfigPermission,
+} from "lib/types";
+import { getPermissionHelper, resolvePermission } from "lib/utils";
 
 const StyledTag = chakra(Tag, {
   baseStyle: {
@@ -27,9 +30,11 @@ export const PermissionChip = ({
 }: PermissionChipProps) => {
   const { address } = useWallet();
 
-  const isAllowed =
-    permissionAddresses.includes(address as HumanAddr) ||
-    instantiatePermission === AccessConfigPermission.EVERYBODY;
+  const isAllowed = resolvePermission(
+    address as HumanAddr,
+    instantiatePermission,
+    permissionAddresses
+  );
 
   const tagBgColor: CSSProperties["backgroundColor"] = isAllowed
     ? "honeydew.darker"