+++ src/lib/hooks/useCodeFilter.ts
@@ -2,7 +2,7 @@ import { useWallet } from "@cosmos-kit/react";
 import { useCallback } from "react";
 
 import type { CodeInfo, HumanAddr } from "lib/types";
-import { AccessConfigPermission } from "lib/types";
+import { resolvePermission } from "lib/utils";
 
 export type PermissionFilterValue =
   | "all"
@@ -13,9 +13,11 @@ export const usePermissionFilter = (filterValue: PermissionFilterValue) => {
   const { address } = useWallet();
   return useCallback(
     ({ instantiatePermission, permissionAddresses }: CodeInfo) => {
-      const isAllowed =
-        permissionAddresses.includes(address as HumanAddr) ||
-        instantiatePermission === AccessConfigPermission.EVERYBODY;
+      const isAllowed = resolvePermission(
+        address as HumanAddr,
+        instantiatePermission,
+        permissionAddresses
+      );
 
       switch (filterValue) {
         case "with-proposal":