+++ src/lib/pages/deploy/index.tsx
@@ -6,27 +6,32 @@ import {
   Button,
   Text,
 } from "@chakra-ui/react";
+import { useWallet } from "@cosmos-kit/react";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
 
-import {
-  useInternalNavigate,
-  useSelectChain,
-  useCurrentNetwork,
-  useWasmConfig,
-} from "lib/app-provider";
+import { useInternalNavigate, useWasmConfig } from "lib/app-provider";
 import { ButtonCard } from "lib/components/ButtonCard";
 import { CustomIcon } from "lib/components/icon";
 import { Stepper } from "lib/components/stepper";
 import WasmPageContainer from "lib/components/WasmPageContainer";
-import { getChainNameByNetwork } from "lib/data";
 import { AmpEvent, AmpTrack } from "lib/services/amplitude";
+import { useGovParams } from "lib/services/proposalService";
+import type { HumanAddr } from "lib/types";
+import { AccessConfigPermission } from "lib/types";
+import { resolvePermission } from "lib/utils";
 
 const Deploy = () => {
-  const { isMainnet } = useCurrentNetwork();
+  const { address } = useWallet();
   const router = useRouter();
   const navigate = useInternalNavigate();
-  const selectChain = useSelectChain();
+  const { data: govParams } = useGovParams();
+  const isAllowed = resolvePermission(
+    address as HumanAddr,
+    govParams?.uploadAccess.permission,
+    govParams?.uploadAccess.addresses
+  );
+
   useWasmConfig({ shouldRedirect: true });
 
   useEffect(() => {
@@ -42,42 +47,54 @@ const Deploy = () => {
       <Heading as="h5" variant="h5" my="48px">
         Select Deploy Option
       </Heading>
-      {isMainnet && (
-        <Alert variant="violet" mb="16px" alignItems="flex-start" gap="1">
-          <CustomIcon
-            name="info-circle-solid"
-            color="violet.ligth"
-            boxSize="20px"
-          />
-          <AlertDescription>
-            Uploading new Wasm files on permissioned chains is coming soon to
-            Celatone. Currently, you can upload codes and instantiate contracts
-            without permission on testnet.
-          </AlertDescription>
-        </Alert>
-      )}
+      {govParams?.uploadAccess.permission !==
+        AccessConfigPermission.EVERYBODY &&
+        (isAllowed ? (
+          <Alert variant="success" mb="16px" alignItems="center" gap={3}>
+            <CustomIcon
+              name="check-circle-solid"
+              color="success.main"
+              boxSize="16px"
+            />
+            <AlertDescription>
+              Your address is included in the whitelist to upload Wasm file.
+            </AlertDescription>
+          </Alert>
+        ) : (
+          <Alert variant="violet" mb="16px" alignItems="center" gap={3}>
+            <CustomIcon
+              name="info-circle"
+              color="violet.light"
+              boxSize="16px"
+            />
+            <AlertDescription>
+              Your selected network is a permissioned chain. Only addresses
+              included in the whitelist are able to upload a new Wasm file.
+            </AlertDescription>
+          </Alert>
+        ))}
       <ButtonCard
         title="Upload new WASM File"
         description={
-          isMainnet ? (
+          !isAllowed ? (
             <Flex fontSize="14px" gap={1}>
               <Text color="text.disabled">
-                Currently available on testnet only.
+                Available for whitelisted addresses only.
               </Text>
               <Text
                 color="honeydew.main"
                 _hover={{ textDecoration: "underline" }}
                 cursor="pointer"
-                onClick={() => selectChain(getChainNameByNetwork("testnet"))}
+                onClick={() => navigate({ pathname: "/proposal/whitelist" })}
               >
-                Switch to testnet
+                Propose to whitelist
               </Text>
             </Flex>
           ) : (
             "Store a new Wasm file on-chain"
           )
         }
-        disabled={isMainnet}
+        disabled={!isAllowed}
         onClick={() => navigate({ pathname: "/upload" })}
         mb="16px"
       />